backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
openai==1.51.0
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
ollama>=0.5.0
